workload-krkn-service-hijacking/pyproject.toml
[tool.poetry]
name = "krkn-service-hijacking"
version = "0.1.0"
description = "A webservice used to simulate HTTP responses in Krkn Scenarios"
authors = ["Tullio Sebastiani <tsebasti@redhat.com>"]
readme = "README.md"
packages = [{include = "app.py", from="./"}]
package-mode=false
license="Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
PyYAML = "^6.0.1"
pytest = "^8.2.0"
coverage = "^7.5.0"
html2text = "^2024.2.26"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
